--- Services/Helpers/Auth.js
@@ -1,7 +1,0 @@
-const passport = require('passport');
-
-const isLoggedIn = () => passport.authenticate('jwt', {
-  session: false,
-});
-
-module.exports = isLoggedIn();
