+++ package.json
@@ -26,6 +26,7 @@
     "eslint-config-airbnb-base": "^12.1.0",
     "eslint-plugin-import": "^2.8.0",
     "express": "^4.16.2",
+    "grpc": "^1.9.1",
     "jsonwebtoken": "^8.1.1",
     "morgan": "^1.9.0",
     "passport": "^0.4.0",
